controller, service 컨벤션..? 어노테이션사용을 확정

+++ src/main/java/com/example/demo11/Dto/BoardDto.java
@@ -9,8 +9,11 @@ public class BoardDto {
     //form에서는 글내용(body), 작성자(writer), 글번호(ID), 댓글번호(reply.Id)notempty
     //본문 글자수 제한을 DB에서 넣지말고 Dto에서...
 
-    private Integer id;
+    private Long id;
     private String content;
     private String writer;
 
+    // BoardDto가 리스트 나오는 페이지  =boardDto = listDto =mainlist, mainpage
+    // Board 클릭하면 Detail
+    // BoardDto에서 글쓰기하면 register
 }
